SE: Windows sockets are not 'int' & not a fd
A Windows socket is a HANDLE which on 64-bit Windows is a 64-bit
type that cannot safely be stored in an 'int'.  Use the SOCKET
abstraction to store sockets on both Windows and UNIX.

A Windows socket is not a CRT file descriptor and cannot be
closed using the CRT close() function.  Use the closesocket()
abstraction to close the socket properly on both Windows and
UNIX.