SE: synchronize updates for Win7, Wow64, debugging
This is a huge patchset that includes a large number of
improvements in the code developed since 2007.

The HANDLE returned from LsaConnectUntrusted() cannot be closed
using CloseHandle().  This was safe to do in 2000, XP and 2003
but it was just an implementation detail that made it so.  Starting
in Vista/2008 and now Win7 the returned handle must be closed with
LsaDeregisterLogonProcess().

Add the ability to report windows errors via the krb5_context
error processing.

Add OutputDebugString debugging messages for builds when
NODEBUG is not defined.

Add support for detecting which versions of WOW64 on XP64 and 2003-64
include working Lsa kerberos support.  Enable the MSLSA cache on
operating systems where it is safe to do so.

Improve memory management.  Avoid buffer overwrites and memory
leaks.