SE: add support for generating Cygwin archives
The generation of Cygwin .a files for each krb5 dll permits
Cygwin applications to link against the KFW binaries which
have access to the MSLSA: and API: credential caches.  The Cygwin
build of the UNIX krb5 distribution only supports FILE: caches.

The KFW gssapi*.dll can also signal NetIdMgr to acquire credentials
or search the available caches for a matching principal.