SE: Support implementation independent krb5 profiles
Check the implementation specific \ProgramData\MIT\Kerberos
as well as the implementation independent \ProgramData\Kerberos.
Return a profile string that includes both.

Always return success even if the profiles cannot be opened.
This permits the default path to be set even if it does not yet
exist.  This is equivalent to how get_from_windows_dir()
behaved when it was the default location.