MSLSA: Do not set initial ticket flag if forwarded
INITIAL and FORWARDED are incompatible flags.  There are cases
where the MIT library is explicitly looking for only an "initial"
ticket but because the LSA is not really a cache but is in fact
a Kerberos ticket request interface, the ticket that is returned
is not an initial ticket but is a valid TGT.   In those cases,
the INITIAL flag is set in the response to improve library
usability.