MSLSA: Restrict KERB_RETRIEVE_TICKET_CACHE_TICKET
The flag KERB_RETRIEVE_TICKET_CACHE_TICKET when applied to a
LSA KerbRetrieveEncodedTicketMessage call obtains a ticket and
adds it to the cached ticket list.  It turns out that the Windows
LSA does not handle the case where more than one ticket for the
same service principal is in the cache at the same time.  This
is particularly true for Ticket Granting Tickets because the LSA
maintains a private cache for "Forwarded" TGTs which can be obtained
as the result of unconstrained delegation.  These TGTs are not meant
for use on the local system but are included in the
KerbQueryTicketCache*Message result set.  If userland queries a
Forwarded TGT when the KERB_RETRIEVE_TICKET_CACHE_TICKET is set,
the Forwarded TGT is moved to the public cache and all subsequent
attempts to obtain a TGT via KerbRetrieveTicketMessage will return
the Forwarded TGT.  This is a problem if the application wishes to
import the TGT into an MIT cache and then renew it because Server 2003
Domain Controllers will not renew a renewable forwarded TGT.

This patchset applies a restriction on the use of
KERB_RETRIEVE_TICKET_CACHE_TICKET so that it will not be used when the
ticket has either the initial or the forwarded flag set.