MSLSA: Filter forwarded TGTs if TGT is initial
When enumerating the KerbQueryTicketCache*Message result set, if a ticket
is an initial TGT, we must skip any TGT with the forwarded flag set
in the response set.  The LSA will obtain a forwarded TGT to use for
unconstrained delegation but this TGT is not useful on the initiator
system.  While it is listed in the response set, it should not be
requested as the LSA does not want to expose it.  If a forwarded TGT is
requested via this interface, a new forwarded TGT will be obtained from
the KDC and it will be returned in place of the initial TGT in the future.
This is a problem because Server 2003 DCs will not renew forwarded TGTs.
There may be other side effects as well.

Filtering out all forwarded TGTs is not an acceptable solution because on an
acceptor system that received delegated credentials there will only be a
forwarded TGT and no initial TGT.  In that case, the forwarded TGT must be
used.